# backend/generated-server/go/model_new_theme_request.go
/*
 * Kotobalize API
 *
 * API for the Kotobalize technical interview preparation app.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type NewThemeRequest struct {

	Title string `json:"title"`

	Description string `json:"description"`

	Category string `json:"category"`

	TimeLimitInSeconds int32 `json:"timeLimitInSeconds"`
}