# backend/generated-server/go/model_update_theme_request.go
/*
 * Kotobalize API
 *
 * API for the Kotobalize technical interview preparation app.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type UpdateThemeRequest struct {

	Title string `json:"title,omitempty"`

	Description string `json:"description,omitempty"`

	Category string `json:"category,omitempty"`

	TimeLimitInSeconds int32 `json:"timeLimitInSeconds,omitempty"`
}