# backend/generated-server/models/model_theme.go
package models

import (
	"time"
)

// Theme model based on openapi.yml
type Theme struct {

	ID int64 `json:"id"`

	Title string `json:"title"`

	Description string `json:"description"`

	Category string `json:"category"`

	TimeLimitInSeconds int `json:"timeLimitInSeconds"`

	CreatedAt time.Time `json:"createdAt"`

	UpdatedAt time.Time `json:"updatedAt"`
}